TEACHING BOARD SCANNER AND WIPER
Board wipers are operated by an electric motor  (back to forth) . 
The electric motor is attached to a worm gear, which transmits the necessary force to a long rod that sets the wiper arms in motion.
The worm gear is able to generate the force required to move the wipers as fast as they need to move
Portable Scanner scans images or pictures quickly: Store JPEG/PDF files within seconds, scan images or pictures quickly, plug and play, no need any software preinstalled.
The arms move the actual wiper blades. The blades are made of cotton cloths and must apply enough pressure to the board in order to remove the dust.
The arms attach to the middle of each wiper blade, and each wiper blade has several arms that distribute the pressure onto the board evenly.